+++ packages/ingest/src/granule.ts
@@ -1,13 +1,13 @@
 import * as crypto from 'crypto';
 import moment from 'moment';
 
-import * as errors from '@cumulus/errors';
-import { s3 } from '@cumulus/aws-client/services';
 import * as S3 from '@cumulus/aws-client/S3';
+import { s3 } from '@cumulus/aws-client/services';
 import * as log from '@cumulus/common/log';
-import { ApiFile, DuplicateHandling, ApiGranule } from '@cumulus/types';
-import { FilePgModel, translatePostgresFileToApiFile, Knex } from '@cumulus/db';
+import { FilePgModel, Knex, translatePostgresFileToApiFile } from '@cumulus/db';
+import * as errors from '@cumulus/errors';
 import { RecordDoesNotExist } from '@cumulus/errors';
+import { ApiFile, DuplicateHandling } from '@cumulus/types';
 
 export interface EventWithDuplicateHandling {
   config: {
@@ -461,23 +461,23 @@ export function unversionFilename(filename: string): string {
 }
 
 /**
- * Generates a unique granule ID by appending a truncated MD5 hash of the granule object.
+ * Generates a unique granule ID by appending a truncated MD5 hash of values from
+ * a producer provided granule object
  *
- * @param granule - The granule object containing granuleId and collectionId.
+ * @param id - An ID associated with the object to be hashed.  Likely the ID
+ * assigned by the granule producer
+ * @param collectionId - The api collection ID (name___version) associated with the granule
  * @param hashLength - The length of the hash to append to the granuleId.
  * @returns - A unique granule ID in the format: granuleId_hash.
  */
-export function generateUniqueGranuleId(granule: ApiGranule, hashLength: number): string {
+export function generateUniqueGranuleId(
+  id: string, collectionId: string, hashLength: number
+): string {
   // use MD5 to generate truncated hash of granule object
   // in the format
-
-  const jsonString = JSON.stringify({
-    collectionId: granule.collectionId,
-    granuleId: granule.granuleId,
-    timestamp: process.hrtime.bigint().toString(),
-  });
-  const hashBuffer = crypto.createHash('md5').update(jsonString).digest();
-  return `${granule.granuleId}_${hashBuffer.toString('base64url').replace(/_/g, '').slice(0, hashLength)}`;
+  const hashString = `${collectionId}_${process.hrtime.bigint().toString()}`;
+  const hashBuffer = crypto.createHash('md5').update(hashString).digest();
+  return `${id}_${hashBuffer.toString('base64url').replace(/_/g, '').slice(0, hashLength)}`;
 }
 
 /**